+++ synapse/SeqSynHandler.cpp
@@ -299,14 +299,14 @@ void SeqSynHandler::fixSynapseOrder()
 
 // Beautiful snippet from Lukasz Wiklendt on StackOverflow. Returns order
 // of entries in a vector.
-template <typename T> vector<size_t> sort_indexes(const vector<T> &v)
+template <typename T> vector<unsigned int> sort_indexes(const vector<T> &v)
 {
     // initialize original index locations
-    vector<size_t> idx(v.size());
+    vector<unsigned int> idx(v.size());
     iota(idx.begin(), idx.end(), 0);
     // sort indexes based on comparing values in v
     sort(idx.begin(), idx.end(),
-         [&v](size_t i1, size_t i2)
+         [&v](unsigned int i1, unsigned int i2)
     {
         return v[i1] < v[i2];
     });
